requirements.txt
gradio
transformers
torch
torchvision
pillow
